Title: 자연수 n개의 합, Time: 128ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240902/자연수 n개의 합/sum-of-n-natural-numbers.java
@@ -3,7 +3,7 @@ import java.util.Scanner;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        long s = sc.nextInt();
+        long s = sc.nextLong();
 
         long l = 1, r = 2000000000;
         long answer = -1;
